+++ pkg/chain/local/local.go
@@ -340,13 +340,13 @@ func (c *localChain) SubmitDKGResult(
 		RequestID:      requestID,
 		MemberIndex:    uint32(participantIndex),
 		GroupPublicKey: resultToPublish.GroupPublicKey[:],
-		BlockNumber:    uint64(currentBlock),
+		BlockNumber:    currentBlock,
 	}
 
 	groupRegistrationEvent := &event.GroupRegistration{
 		GroupPublicKey: resultToPublish.GroupPublicKey[:],
 		RequestID:      requestID,
-		BlockNumber:    uint64(currentBlock),
+		BlockNumber:    currentBlock,
 	}
 
 	c.handlerMutex.Lock()
